+++ ios/history.md
@@ -9,6 +9,7 @@
 * Added support for keypress error feedback (#257)
 * Fixed ongoing issues with keyboard rotation and sizing, including the iPhone X notch. (#444) (#1045)
 * Fixed keyboard display/overlap of "Getting Started" info panel, added keyboard hide/display API functions. (#1084)
+* Fixes issues with keyboard keycap scaling and diacritic display. (#1070)
 
 ## 2018-08-02 10.0.208 stable 
 * Fixed OSK layout problems (and possible crash) on iOS 11 on certain hardware (#1089, #1159)